💄 refactor(chat): forward via dialog instead of hover popover
Revert the Forward affordance from a hover popover back to a click-triggered
dialog (search + multi-select agents + note + submit). Esc still closes the
dialog first, then exits selection mode on a second press.
